+++ src/app/shared/PersonName.svelte
@@ -7,19 +7,24 @@
     session,
     follows,
     getFollowsWhoFollow,
+    possibleImposters,
   } from "src/engine"
 
   export let pubkey
 
   const person = derivePerson(pubkey)
+
   const following = deriveFollowing(pubkey)
+
   const followCount = follows.derived(pks => {
-    if ($following || pubkey === $session.pubkey) {
+    if (!$session || $following || pubkey === $session.pubkey) {
       return null
     }
 
     return getFollowsWhoFollow($session.pubkey, pubkey).length
   })
+
+  const mayBeImposter = possibleImposters.derived(pubkeys => pubkeys.has(pubkey))
 </script>
 
 <div class={cx("flex items-center gap-1", $$props.class)}>
@@ -28,6 +33,10 @@
     {#if $following}
       <span class="px-2 py-1 text-xs">
         <i class="fa fa-user-check pl-px text-accent" />
+      </span>
+    {:else if $mayBeImposter}
+      <span class="px-2 py-1 text-xs">
+        <i class="fa fa-warning text-warning" />
       </span>
     {:else if $followCount}
       <span class="px-2 py-1 text-xs">